All tabs should be displayed in a single route
Don't use nested routes. With nested routes for each tab switch the expensive search query is executed again.
With a single route we execute the search query once for all entities and the user can switch tabs with low latency.
We also can add the number of found entities for the search beside the tab labels.
TODO:
- Investigate short query bug -> prisma endless loop
- Search input field via searchParams
- Pagination
- Show number of results in labels
- Build tabs remix style (Accessibility)
- Authorization (incl. publicFields)
